backend/app/utils/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beanie==1.26.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
fastapi==0.114.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.0
langchain-openai==0.2.0
langsmith==0.1.121
lazy-model==0.2.0
motor==3.5.1
openai==1.45.0
orjson==3.10.7
packaging==24.1
pydantic==2.9.1
pydantic_core==2.23.3
pymongo==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
sniffio==1.3.1
starlette==0.38.5
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
